requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
python-multipart
pillow
numpy
